pyproject.toml
[tool.pdm.build]
includes = []
[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"


[project]
name = "KuoHuBit"
version = "114514.1919810"
description = "A QQ Bot"
authors = [
    { name = "daizihan233", email = "y066ydg5@duck.com" },
]
dependencies = [
    "graia-ariadne[full]>=0.11.7",
    "arclet-alconna<2.0.0, >=1.8.1",
    "arclet-alconna-tools<0.8.0, >=0.7.0",
    "tarina>=0.4.2",
    "nepattern<0.7.0, >=0.6.0",
    "graia-broadcast>=0.18.2",
    "launart>=0.5.0",
    "pyyaml<7.0,>=6.0",
    "requests-cache<1.0.0,>=0.9.7",
    "redis<5.0.0,>=4.4.0",
    "pymysql<2.0.0,>=1.0.2",
    "requests<3.0.0,>=2.28.1",
    "aiohttp[speedups]<4.0.0,>=3.8.3",
    "aiomysql<1.0.0,>=0.1.1",
    "opencc-python-reimplemented<1.0.0,>=0.1.6",
    "jieba<1.0.0,>=0.42.1",
    "portalocker>=2.7.0",
    "numpy>=1.24.2",
    "uvicorn>=0.21.1",
    "fastapi>=0.95.1",
    "rich>=13.4.2",
    "tencentcloud-sdk-python-tms>=3.0.967",
    "tencentcloud-sdk-python-ims>=3.0.967",
    "pypinyin>=0.49.0",
    "setuptools>=69.0.2",
    "pip>=23.3.1",
    "g4f[all]>=0.2.1.3",
    "openai>=1.12.0",
    "tiktoken>=0.6.0",
    "arclet-alconna-graia>=0.14.0",
    "arclet-alconna-ariadne>=0.14.0",
    "graiax-shortcut>=0.3.0",
    "tokencost>=0.1.11",
]
requires-python = ">=3.10,<3.13"
license = { text = "AGPL-3.0" }
readme = "README.md"

